pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "PiWS2811Controller"
description = "Plugin-based controller for Neopixel LEDs."
version = "0.0.1"
authors = [{name = "Iamaprogrammer"}]
dependencies = [
    "adafruit-circuitpython-neopixel"
]